Create README.md
This game was made in a group, I received help from 2 other students. 
I was responsible for the main class 'MurderMadness', where all the game logic and initialisation happens.
We are in the process of making a GUI for this, to make the game more user friendly
Currently the game has some issues such as it doesn't tell you which house is which until you entre the houses. 
The game also doesn't have a instructions yet, it expects you to have a basic knowledge of Cluedo to play it.